# api/openapi-spec.yaml
openapi: 3.0.0
info:
  title: Images DB
  description: |
    This is the Images DB API based on the OpenAPI 3.0.0 specification.  
    Users are able to upload their images on the server and retrieve it.
  contact:
    name: Firas
    email: firasfj17@gmail.com
  version: 1.0.0
servers:
  - url: http://localhost:4000/v1
...
security:
  - jwt_auth: []
paths:
...
  /users:
    get:
      security:
        - jwt_auth: []
      tags:
        - users
      summary: Retrieve a batch of users
      description: Retrieve a batch of users with pagination and sorting (For admin)
      operationId: fetchBatchOfUsers
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
          description: Number of users to return
          example: 10
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
          description: Number of users to skip
          example: 0
        - in: query
          name: sortBy
          schema:
            type: string
          description: Field to sort by
          example: id
        - in: query
          name: sortOrder
          schema:
            type: string
          description: Sort order (ASC or DESC)
          example: ASC
        - in: query
          name: showDeleted
          schema:
            type: string
          description: Whether to include deleted items
          example: true
        - in: query
          name: searchQuery
          schema:
            type: string
          description: Search query
          example: thomas
        - in: query
          name: searchColumn
          schema:
            type: string
          description: Column to search in
          example: userName
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Successfully fetched users
                  fetched:
                    type: integer
                    example: 10
                  data:
                    $ref: "#/components/schemas/UserArray"
                  links:
                    $ref: "#/components/schemas/HATEOASLinkArray"
                  totalUsers:
                    type: integer
                    example: 10
                  totalNeededUsers:
                    type: integer
                    example: 10
        "400":
          $ref: "#/components/responses/InvalidInputError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          description: Only the admin can use this route
        "404":
          description: No users found
        "500":
          $ref: "#/components/responses/ServerError"

    post:
      security:
        - jwt_auth: []
      tags:
        - users
      summary: Create a new user
      description: Create a new user (For admin)
      operationId: createUser
      requestBody:
        description: Create a new user
        content:
          application/json:
            schema:
              type: object
              required:
                - userName
                - password
              properties:
                userName:
                  type: string
                  maxLength: 15
                password:
                  type: string
                  maxLength: 30
      responses:
        "201":
          description: Successfully created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User created successfully
                  userId:
                    type: integer
                    example: 301
                  links:
                    $ref: "#/components/schemas/HATEOASLinkArray"
        "400":
          $ref: "#/components/responses/InvalidInputError"
        "403":
          description: Only the admin can create users using this route!
        "409":
          description: Username already exists
        "500":
          $ref: "#/components/responses/ServerError"
  /users/{userId}:
    get:
      security:
        - jwt_auth: []
      tags:
        - users
      summary: Retrieve user by ID
      description: Retrieve a user by their ID (For admin)
      operationId: getUserById
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
          description: ID of the user to retrieve
          example: 10

      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Found user!
                  data:
                    $ref: "#/components/schemas/User"
                  links:
                    $ref: "#/components/schemas/HATEOASLinkArray"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "400":
          $ref: "#/components/responses/InvalidInputError"
        "403":
          description: Only the admin can use this route
        "404":
          description: User not found
        "500":
          $ref: "#/components/responses/ServerError"

    patch:
      security:
        - jwt_auth: []
      tags:
        - users
      summary: Partially update a user by ID
      description: Partially update a user's information by specifying the fields to update (For admin)
      operationId: partiallyUpdateUserById
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
          description: ID of the user to update
          example: 10
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  example: 10
                userName:
                  type: string
                  example: katie
                isAdmin:
                  type: boolean
                  example: false
                createdAt:
                  type: string
                  oneOf:
                    - format: date
                    - format: date-time
                  example: 2023-06-19
                updatedAt:
                  type: string
                  oneOf:
                    - format: date
                    - format: date-time
                  example: 2023-06-19
                destroyTime:
                  type: string
                  oneOf:
                    - format: date
                    - format: date-time
                  nullable: true
                  example: null
      responses:
        "200":
          description: User updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  links:
                    $ref: "#/components/schemas/HATEOASLinkArray"
        "400":
          $ref: "#/components/responses/InvalidInputError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          description: Only the admin can update the users
        "404":
          description: User not found
        "500":
          $ref: "#/components/responses/ServerError"

    put:
      security:
        - jwt_auth: []
      tags:
        - users
      summary: Update a user by ID
      description: Update a user's information by providing all fields (For admin)
      operationId: updateUserById
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
          description: ID of the user to update

      requestBody:
        required: true
        $ref: "#/components/schemas/User"
      responses:
        "200":
          description: User updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  links:
                    $ref: "#/components/schemas/HATEOASLinkArray"
        "400":
          $ref: "#/components/responses/InvalidInputError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          description: Only the admin can update the users
        "404":
          description: User not found
        "500":
          $ref: "#/components/responses/ServerError"

    delete:
      security:
        - jwt_auth: []
      tags:
        - users
      summary: Delete a user by ID
      description: Delete a user by providing their ID (For admin)
      operationId: deleteUserById
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
          description: ID of the user to delete
          example: 10
      responses:
        "200":
          description: User deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User deleted successfully!
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          description: Only the admin can delete the other users
        "404":
          description: User not found
        "500":
          $ref: "#/components/responses/ServerError"
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 10
        userName:
          type: string
          example: katie
        isAdmin:
          type: boolean
          example: false
        createdAt:
          type: string
          oneOf:
            - format: date
            - format: date-time
          example: 2023-06-19
        updatedAt:
          type: string
          oneOf:
            - format: date
            - format: date-time
          example: 2023-06-19
        destroyTime:
          type: string
          oneOf:
            - format: date
            - format: date-time
          nullable: true
          example: null
      required:
        - id
        - userName
...
    UserArray:
      description: Array of users
      type: array
      items:
        $ref: "#/components/schemas/User"
...
    HATEOASLink:
      type: object
      properties:
        rel:
          type: string
        method:
          type: string
        href:
          type: string
        description:
          type: boolean
    HATEOASLinkArray:
      description: Array of HATEOAS links
      type: array
      items:
        $ref: "#/components/schemas/HATEOASLink"
  responses:
    UnauthorizedError:
      description: Token not provided or invalid
    InvalidInputError:
      description: Invalid input
    ServerError:
      description: Internal server error occurred.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              details:
                type: string
  securitySchemes:
    jwt_auth:
      type: apiKey
      in: cookie
      name: jwt